set icons via pixbuf rather than saving to temporary files
Needs a workaround for setting the default icon (when a tune has no associated
art) since setting a gnome icon does not override the previously set pixbuf
icon. But in finding that workaround I also found how to display that default
icon at the same size as the album art.